Implement fractional ratings (0.0-5.0)
- Update database schema to use REAL for score
- Update Rust models and handlers for f64 support
- Update UI with SVG stars and touch-friendly 3-step toggle (Full/Half/Clear)
- Add integration tests for fractional ratings

+++ src/models.rs
@@ -112,5 +112,5 @@ pub struct Rating {
     pub id: i64,
     pub recipe_id: i64,
     pub rater_name: String,
-    pub score: i64,
+    pub score: f64,
 }
